+++ src/shapes/MeshTop_t.cpp
@@ -12,7 +12,7 @@ APTracer::Shapes::MeshTop_t::MeshTop_t(APTracer::Entities::Material_t *material,
 
 APTracer::Shapes::MeshTop_t::~MeshTop_t(){
     if (triangles_ != nullptr){
-        for (unsigned int i = 0; i < n_tris_; i++){
+        for (size_t i = 0; i < n_tris_; i++){
             if (triangles_[i] != nullptr){
                 delete triangles_[i];
             }
@@ -22,7 +22,7 @@ APTracer::Shapes::MeshTop_t::~MeshTop_t(){
 }
 
 void APTracer::Shapes::MeshTop_t::update(){
-    for (unsigned int i = 0; i < n_tris_; i++){
+    for (size_t i = 0; i < n_tris_; i++){
         triangles_[i]->update();
     }
 }
@@ -49,7 +49,7 @@ void APTracer::Shapes::MeshTop_t::normal(const APTracer::Entities::Ray_t &ray, c
 
 Vec3f APTracer::Shapes::MeshTop_t::mincoord() const {
     Vec3f coord = Vec3f(std::numeric_limits<double>::infinity());
-    for (unsigned int i = 0; i < n_tris_; i++){
+    for (size_t i = 0; i < n_tris_; i++){
         coord.min(triangles_[i]->mincoord());
     }    
     return coord;
@@ -57,7 +57,7 @@ Vec3f APTracer::Shapes::MeshTop_t::mincoord() const {
 
 Vec3f APTracer::Shapes::MeshTop_t::maxcoord() const {
     Vec3f coord = Vec3f(-std::numeric_limits<double>::infinity());
-    for (unsigned int i = 0; i < n_tris_; i++){
+    for (size_t i = 0; i < n_tris_; i++){
         coord.max(triangles_[i]->maxcoord());
     }
     return coord;
