+++ src/shapes/Mesh_t.cpp
@@ -21,14 +21,14 @@ APTracer::Shapes::Mesh_t::~Mesh_t(){}
 
 void APTracer::Shapes::Mesh_t::createTriangles(){    
     triangles_ = new APTracer::Entities::Shape_t*[n_tris_];
-    for (unsigned int i = 0; i < n_tris_; i++){
+    for (size_t i = 0; i < n_tris_; i++){
         triangles_[i] = new APTracer::Shapes::TriangleMesh_t(material_, transformation_, geom_, i);
     }
 }
 
 void APTracer::Shapes::Mesh_t::createTriangles(APTracer::Entities::MaterialMap_t *materialmap){
     triangles_ = new APTracer::Entities::Shape_t*[n_tris_];
-    for (unsigned int i = 0; i < n_tris_; i++){
+    for (size_t i = 0; i < n_tris_; i++){
         triangles_[i] = new APTracer::Shapes::TriangleMesh_t(materialmap->getMaterial(geom_->mat_[i]), transformation_, geom_, i);
     }
 }
